pyproject.toml
[tool.poetry]
name = "python-auth-api"
version = "0.1.0"
description = "An example project for python API best practices with FastAPI and MongoDB"
authors = ["Kieran Wynn"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
uvicorn = "^0.13.4"
python-decouple = "^3.4"
pydantic = {version = "^1.8.2", extras = ["email"]}
python-multipart = "^0.0.5"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-jose = {version = "^3.3.0", extras = ["cryptography"]}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
